requirements.txt
# Static Site Generator Dependencies

# Core dependencies
pyyaml>=6.0
jinja2>=3.1
markdown-it-py>=3.0
mdit-py-plugins>=0.4
pygments>=2.17
bibtexparser>=1.4.0
python-slugify>=8.0
beautifulsoup4>=4.12.0

# Optional for live reload (v1.1)
# watchdog>=4.0
